+++ hc_dpki/src/lib.rs
@@ -4,10 +4,11 @@ extern crate holochain_sodium;
 #[macro_use]
 extern crate arrayref;
 extern crate base64;
-extern crate rustc_serialize;
-
 extern crate bip39;
 extern crate boolinator;
+extern crate rustc_serialize;
+#[macro_use]
+extern crate serde;
 
 pub mod bundle;
 pub mod error;