+++ hdk-rust/src/macros.rs
@@ -195,7 +195,7 @@ macro_rules! define_zome {
                             outputs: | $( $output_param_name:ident : $output_param_type:ty ),* |,
                             handler: $handler_path:path
                         }
-            )+
+            )*
         ]
 
         capabilities : {
@@ -295,7 +295,7 @@ macro_rules! define_zome {
                         capability.functions = vec![
                             $(
                                 stringify!($cap_fn).into()
-                            ),+
+                            ),*
                         ];
 
                         cap_map.insert(stringify!($cap).into(), capability);
@@ -332,7 +332,7 @@ macro_rules! define_zome {
                                         ),*
                                     ]
                                 }
-                    ),+
+                    ),*
 
                 ]
             };
@@ -375,6 +375,6 @@ macro_rules! define_zome {
                         $crate::global_fns::write_json(execute(input))
                     ).into()
                 }
-        )+
+        )*
     };
 }